Another version of the laplacian interpolator.
Both versions of the function regionfill3D are clones
of the matlab function regionfill, but for 3D.

The driver function is speedimage, which needs to be given a
distance transform image.

The old regionfill3D ends up creating an n x n matrix (sparse),
where n is the number of voxels in the 3D image.
The new regionfill3D creates an m x m matrix (sparse),
where m is the number of voxels in the mask, after dilating by 1.

Setting up the matrix structure is messier in the new regionfill3D, but the
final matrix will be much smaller. Not yet sure how much practical difference
this will make.