Speed image from distance transform, quickly.
A tophat filtering of the distance transform, followed
by blurring and including a constant background. The
script currently doesn't handle all the options available
to the R function (blurring size etc).

The tophat detects ridge lines in the distance transform,
which is similar to the skeleton but without the cost.

The constant background means that deviations from the
ridgelines are penalized equally, which may be a bad thing.

The chance of leaving the ridgeline depends a lot on the background
speed, which is currently set very low.